+++ package.json
@@ -15,5 +15,9 @@
   "bugs": {
     "url": "https://github.com/webmarkyn/weather-app/issues"
   },
-  "homepage": "https://github.com/webmarkyn/weather-app#readme"
+  "homepage": "https://github.com/webmarkyn/weather-app#readme",
+  "devDependencies": {
+    "webpack": "^4.41.5",
+    "webpack-cli": "^3.3.10"
+  }
 }
