+++ src/colors.js
@@ -1,0 +1,28 @@
+const weatherColor = {
+  Clear: {
+    color: "#34e7e4",
+    icon: "🔆"
+  },
+  Snow: {
+    color: "#f5f6fa",
+    icon: "❄️"
+  },
+  Fog: {
+    color: "#d2dae2",
+    icon: "🌁"
+  },
+  Mist: {
+    color: "#d2dae2",
+    icon: "🌁"
+  },
+  Rain: {
+    color: "#808e9b",
+    icon: "☂️"
+  },
+  Clouds: {
+    color: "#dcdde1",
+    icon: "☁️"
+  }
+};
+
+export default weatherColor
