+++ webpack.config.js
@@ -1,0 +1,12 @@
+const path = require('path');
+
+module.exports = {
+    entry: "./src/index.js",
+    output: {
+        path: path.join(__dirname, 'dist'),
+        filename: 'main.js'
+    },
+    devServer: {
+      contentBase: './dist',
+    },
+};
